# fix!: better cache memory usage