* Migrate from archived https://github.com/kevburnsjr/skipfilter to https://github.com/dunglas/skipfilter
* Migrate from `golang-lru` to Otter for better performance
* Migrate from `hash/fnv` to `github.com/cespare/xxhash/v2`
* Hashers pool
* Bump deps

Should help for #1024