@naxo8628 2025-10-01
Hi @dunglas, thank you so much for looking into this and creating the PR! I'd be happy to test these changes in my environment. Is there a pre-built Docker image available for this pull request that I could use?

I'm new to Go and I'm trying to compile the project and then build a Docker image.

First, I compiled the binary with:
`go build -o mercure ./cmd/mercure/`

Then, I built the Docker image using:
`docker buildx build --platform linux/amd64 -t myrepo.../mercure:test-cache . --push`

However, when the pod starts up, it fails with the following error:
`exec /usr/bin/caddy: exec format error`

Any help would be appreciated. Thanks!