@naxo8628 2025-10-01
@dunglas  Solved! I've never used goreleaser.

I've deployed this version to the staging environment, and it will go to production tomorrow. I'll provide you with performance metrics after that.