@dunglas 2025-10-02
Here is a benchmark I ran locally:

```console
SUB_TEST_CONCURRENCY=500 SUB_TEST_TOPICS=100 SUB_TEST_MATCHPCT=70 go test -benchmem -run='^$' -count 6 -bench '^BenchmarkLocalTransport$' > old.txt
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/dunglas/mercure
cpu: Apple M1 Pro
                                                         │       old.txt       │               new.txt               │
                                                         │       sec/op        │       sec/op        vs base         │
LocalTransport/100-topics:500-concurrency:70-matchpct-10   107.96µ ± 16687715%   96.89µ ± 30112408%  ~ (p=0.065 n=6)

                                                         │       old.txt       │                   new.txt                   │
                                                         │        B/op         │         B/op          vs base               │
LocalTransport/100-topics:500-concurrency:70-matchpct-10   37.60Ki ± 74880960%   22.27Ki ± 209187700%  -40.79% (p=0.026 n=6)

                                                         │       old.txt       │                  new.txt                   │
                                                         │      allocs/op      │      allocs/op       vs base               │
LocalTransport/100-topics:500-concurrency:70-matchpct-10   31.00 ± 1100182848%   22.00 ± 2566207586%  -29.03% (p=0.015 n=6)
```