@dunglas 2025-10-02
Thanks @naxo8628. This indeed looks way better. I'm still interested in a profile, if possible.

Thank you.