@naxo8628 2025-10-03
@dunglas I've just deployed to production with debugging enabled, as I wanted to wait at least 24 hours before the next deployment. I will send performance profiles within the next 48 hours.

In any case, although the last deploy improved the initial memory spike, it seems the memory fills up again after a few hours.


<img width="584" height="262" alt="Captura de pantalla 2025-10-03 a las 10 56 48" src="https://github.com/user-attachments/assets/5b618177-2ca4-4c24-83d3-a559617b1ccb" />